+++ CHANGELOG.md
@@ -21,13 +21,14 @@ and this project adheres _loosely_ to [Semantic Versioning](https://semver.org/s
 
 ### Added
 
-- Add following M1 extensions axes methods: Ancestor, PrecedingSiblings, FollowingSiblings
+- Add following M1 extensions axes methods: `Ancestor`, `PrecedingSiblings`, `FollowingSiblings`.
 - Generator adds descriptions to generated classes from model annotations.
 - Language (de-)Serialization can handle annotations on languages.
 - Generate interface for each factory; factory implementation methods are `virtual` now.
 - Add utilities class `ReferenceUtils` (in namespace `LionWeb.Core.Utilities`) to deal with references:
 	- `ReferenceValues(<nodes>)` finds all references within all the given `<nodes>`.
 	- `FindIncomingReferences(<targetNode(s)>, <nodes>)` finds all references _to_ (all) the `<targetNode(s)>` within the given `<nodes>`.
+    - `ReferencesToOutOfScopeNodes(<nodes>)` finds all references to nodes that are not withing the given `<nodes>`.
 
 ### Fixed